+++ crates/holochain/src/core/ribosome/host_fn/send_remote_signal.rs
@@ -142,7 +142,7 @@ mod tests {
                 api.emit_signal(AppSignal::new(signal)).map_err(Into::into)
             })
             .function("init", move |api, ()| {
-                let mut fns = BTreeSet::new();
+                let mut fns = HashSet::new();
                 fns.insert((api.zome_info(()).unwrap().name, "recv_remote_signal".into()));
                 let functions = GrantedFunctions::Listed(fns);
                 let cap_grant_entry = CapGrantEntry {