+++ crates/test_utils/wasm/wasm_workspace/coordinator_zome/src/lib.rs
@@ -1,3 +1,4 @@
+use std::collections::HashSet;
 use hdk::prelude::*;
 use integrity_zome::EntryTypes;
 use integrity_zome::Msg;
@@ -89,7 +90,7 @@ fn get_activity(
 #[hdk_extern]
 fn init(_: ()) -> ExternResult<InitCallbackResult> {
     // grant unrestricted access to accept_cap_claim so other agents can send us claims
-    let mut fns = BTreeSet::new();
+    let mut fns = HashSet::new();
     fns.insert((zome_info()?.name, "create_entry".into()));
     let functions = GrantedFunctions::Listed(fns);
     create_cap_grant(CapGrantEntry {